+++ changes.md
@@ -9,9 +9,11 @@
 * Added door package
   * Added VDoor
   * Added ArchDoor
-* Updated README.md
 * Added build return annotations for tower classes.
 * Modified tower TowerDoor to have better custom outline support.
+* Wrote Documentation for Arch, Door, House, and House Wall
+* Added various missing type annotations when I spotted them.
+* Updated README.md
 
 ## 2.0.0
 * Updated cqterrain to version 3.4.1